+++ R/RcppExports.R
@@ -5,8 +5,8 @@ ffTestStatistic <- function(S1, S2, method) {
     .Call(`_fasano_franceschini_test_ffTestStatistic`, S1, S2, method)
 }
 
-permutationTestSeeded <- function(S1, S2, nPermute, verbose, seed, method) {
-    .Call(`_fasano_franceschini_test_permutationTestSeeded`, S1, S2, nPermute, verbose, seed, method)
+permutationTestSeeded <- function(S1, S2, nPermute, verbose, method, seed) {
+    .Call(`_fasano_franceschini_test_permutationTestSeeded`, S1, S2, nPermute, verbose, method, seed)
 }
 
 permutationTest <- function(S1, S2, nPermute, verbose, method) {
@@ -15,5 +15,9 @@ permutationTest <- function(S1, S2, nPermute, verbose, method) {
 
 permutationTestParallel <- function(S1, S2, nPermute, method) {
     .Call(`_fasano_franceschini_test_permutationTestParallel`, S1, S2, nPermute, method)
+}
+
+permutationTestParallelSeeded <- function(S1, S2, nPermute, method, seed) {
+    .Call(`_fasano_franceschini_test_permutationTestParallelSeeded`, S1, S2, nPermute, method, seed)
 }