+++ NEWS.md
@@ -1,6 +1,7 @@
 # fasano.franceschini.test 2.1.0
 * The `'o'` (optimize) method has been removed as the benchmarking itself proved to be prohibitively slow. Instead, the default behavior now is to infer which method is likely faster based on the sample sizes and dimension of the data.
 * The `cores` argument has been removed. Use `threads` instead.
+* Several bottlenecks in the code were cleared.
 
 # fasano.franceschini.test 2.0.1
 * Minor speedups introduced, particularly for the brute force method.